requirements.txt
pyarrow==16.0.0
python-dotenv
google-cloud-storage
pandas
requests